backend/python/requirements-fixed.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# Authentication & JWT
djangorestframework-simplejwt==5.3.0

# WebSocket & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4

# File Storage & Media
Pillow==10.0.1
whitenoise==6.6.0
boto3==1.42.91

# Data Processing (compatible versions)
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.0.3

# ML/AI (compatible versions)
tensorflow==2.15.0
tensorflow-intel==2.15.0

# Computer Vision (compatible with numpy 1.26.4)
opencv-python==4.8.1.78

# HTTP & API
requests==2.31.0
urllib3==2.0.7

# Security & Crypto
cryptography==46.0.7
pyjwt==2.12.1
python3-openid==3.2.0

# Date & Time
python-dateutil==2.9.0.post0
pytz==2026.1

# Development & Testing
gunicorn==21.2.0

# Optional: Comment out if not needed
# google-auth==2.25.2
# google-auth-oauthlib==1.2.0
# google-auth-httplib2==0.2.0
